+++ src/test/java/org/folio/circulation/api/APITestSuite.java
@@ -35,7 +35,8 @@ import static org.hamcrest.core.Is.is;
 @Suite.SuiteClasses({
   LoanAPITests.class,
   LoanRulesAPITests.class,
-  LoanRulesEngineAPITests.class
+  LoanRulesEngineAPITests.class,
+  RequestAPITests.class
 })
 public class APITestSuite {
   private static final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());