+++ src/test/java/org/folio/circulation/api/RequestAPITests.java
@@ -1,0 +1,192 @@
+package org.folio.circulation.api;
+
+import io.vertx.core.json.JsonObject;
+import org.folio.circulation.support.http.client.OkapiHttpClient;
+import org.folio.circulation.support.http.client.Response;
+import org.folio.circulation.support.http.client.ResponseHandler;
+import org.junit.Before;
+import org.junit.Test;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.io.UnsupportedEncodingException;
+import java.lang.invoke.MethodHandles;
+import java.net.HttpURLConnection;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.util.UUID;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
+
+import static org.hamcrest.core.Is.is;
+import static org.hamcrest.junit.MatcherAssert.assertThat;
+
+public class RequestAPITests {
+  private static final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
+
+  OkapiHttpClient client = APITestSuite.createClient(exception -> {
+    log.error("Request to circulation module failed:", exception);
+  });
+
+  @Before
+  public void beforeEach()
+    throws MalformedURLException,
+    InterruptedException,
+    ExecutionException,
+    TimeoutException {
+
+    deleteAllLoans();
+    deleteAllItems();
+  }
+
+  @Test
+  public void createRequestIsNotImplemented()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException,
+    UnsupportedEncodingException {
+
+    CompletableFuture<Response> postCompleted = new CompletableFuture<>();
+
+    client.post(requestsUrl(), new JsonObject(), ResponseHandler.any(postCompleted));
+
+    Response postResponse = postCompleted.get(5, TimeUnit.SECONDS);
+
+    assertThat(postResponse.getStatusCode(), is(HttpURLConnection.HTTP_NOT_IMPLEMENTED));
+  }
+
+  @Test
+  public void getRequestByIdIsNotImplemented()
+    throws MalformedURLException,
+    InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    UnsupportedEncodingException {
+
+    CompletableFuture<Response> getCompleted = new CompletableFuture<>();
+
+    client.get(requestsUrl(String.format("/%s", UUID.randomUUID())),
+      ResponseHandler.any(getCompleted));
+
+    Response getResponse = getCompleted.get(5, TimeUnit.SECONDS);
+
+    assertThat(getResponse.getStatusCode(), is(HttpURLConnection.HTTP_NOT_IMPLEMENTED));
+  }
+
+  @Test
+  public void getAllRequestsIsNotImplemented()
+    throws MalformedURLException,
+    InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    UnsupportedEncodingException {
+
+    CompletableFuture<Response> getCompleted = new CompletableFuture<>();
+
+    client.get(requestsUrl(), ResponseHandler.any(getCompleted));
+
+    Response getResponse = getCompleted.get(5, TimeUnit.SECONDS);
+
+    assertThat(getResponse.getStatusCode(), is(HttpURLConnection.HTTP_NOT_IMPLEMENTED));
+  }
+
+  @Test
+  public void replaceRequestIsNotImplemented()
+    throws InterruptedException,
+    MalformedURLException,
+    TimeoutException,
+    ExecutionException,
+    UnsupportedEncodingException {
+
+    CompletableFuture<Response> putCompleted = new CompletableFuture<>();
+
+    client.put(requestsUrl(String.format("/%s", UUID.randomUUID())),
+      new JsonObject(), ResponseHandler.any(putCompleted));
+
+    Response putResponse = putCompleted.get(5, TimeUnit.SECONDS);
+
+    assertThat(putResponse.getStatusCode(), is(HttpURLConnection.HTTP_NOT_IMPLEMENTED));
+
+  }
+
+  @Test
+  public void deletingAllRequestsIsNotImplemented()
+    throws InterruptedException,
+    MalformedURLException,
+    TimeoutException,
+    ExecutionException,
+    UnsupportedEncodingException {
+
+    CompletableFuture<Response> deleteCompleted = new CompletableFuture<>();
+
+    client.delete(requestsUrl(),
+      ResponseHandler.any(deleteCompleted));
+
+    Response deleteResponse = deleteCompleted.get(5, TimeUnit.SECONDS);
+
+    assertThat(deleteResponse.getStatusCode(), is(HttpURLConnection.HTTP_NOT_IMPLEMENTED));
+  }
+
+  @Test
+  public void deletingARequestIsNotImplemented()
+    throws InterruptedException,
+    MalformedURLException,
+    TimeoutException,
+    ExecutionException,
+    UnsupportedEncodingException {
+
+    CompletableFuture<Response> deleteCompleted = new CompletableFuture<>();
+
+    client.delete(requestsUrl(String.format("/%s", UUID.randomUUID())),
+      ResponseHandler.any(deleteCompleted));
+
+    Response deleteResponse = deleteCompleted.get(5, TimeUnit.SECONDS);
+
+    assertThat(deleteResponse.getStatusCode(), is(HttpURLConnection.HTTP_NOT_IMPLEMENTED));
+  }
+
+  private static URL itemsUrl(String subPath)
+    throws MalformedURLException {
+
+    return APITestSuite.viaOkapiModuleUrl("/item-storage/items" + subPath);
+  }
+
+  private static URL loansUrl(String subPath)
+    throws MalformedURLException {
+
+    return APITestSuite.viaOkapiModuleUrl("/loan-storage/loans" + subPath);
+  }
+
+  private static URL requestsUrl()
+    throws MalformedURLException {
+
+    return requestsUrl("");
+  }
+
+  private static URL requestsUrl(String subPath)
+    throws MalformedURLException {
+
+    return APITestSuite.circulationModuleUrl("/circulation/requests" + subPath);
+  }
+
+  private void deleteAllLoans()
+    throws MalformedURLException,
+    InterruptedException,
+    ExecutionException,
+    TimeoutException {
+
+    APITestSuite.deleteAll(loansUrl(""));
+  }
+
+  private void deleteAllItems()
+    throws MalformedURLException,
+    InterruptedException,
+    ExecutionException,
+    TimeoutException {
+
+    APITestSuite.deleteAll(itemsUrl(""));
+  }
+}